[Feat] BuildingInfoScreen 사이의 navigation 연결하기 #30

+++ app/src/main/java/com/ganaljigi/kubf/navigation/Routes.kt
@@ -21,7 +21,7 @@ sealed interface Routes {
     data object Notice : Routes
 
     @Serializable
-    data class BuildingInfo(val number: Int) : Routes
+    data class BuildingInfo(val number: Long) : Routes
 
     @Serializable
     data class RoomInfo(
